Move body constructors to helpers module
