+++ src/core/http-body/helpers.ts
@@ -1,0 +1,13 @@
+import { HttpBody } from "../http";
+import { StringBody } from "./string-body";
+
+export async function asJson<T>(body: HttpBody): Promise<T> {
+  const bodyToString = await body.asString();
+
+  return JSON.parse(bodyToString);
+}
+
+export function jsonBody(obj: any) {
+  const objStr = JSON.stringify(obj);
+  return new StringBody(objStr);
+}
